Bugfix: JMAK model estimation
Parameters are now estimated using time points for the intercept
instead of the data points (that was incorrect). Added extra
treating of data points that are exactly 0 or exactly max_frac
as this results in invalid values in the double log parameter
estimation. Changed plotting to make sure that they are readable.