Update: biased weighted linear regression
To deal with difficult genetic regions, use biased weighted
linear regression. Biased because a requirement is set that
all time steps need to differ by at least 1%-point increase.
Weighted because if the data point at 20 min is 0, we incorporate
a weight for implementing confidence in different areas.
For difficult regions, both is applied